Suisan-neki

+++ client/src/lib/audioAnalyzer.ts
@@ -429,7 +429,7 @@ export class AudioAnalyzer {
     if (!this.isRunning || !this.analyser || !this.dataArray) return;
     
     // 時間データを取得
-    this.analyser.getFloatTimeDomainData(this.dataArray);
+    this.analyser.getFloatTimeDomainData(this.dataArray as Float32Array);
     
     // RMSを計算
     const rms = this.calculateRMS(this.dataArray);
